custom hook으로 수정
