+++ src/components/Roof.tsx
@@ -1,75 +1,26 @@
+import { useCanvas } from '@/hooks/useCanvas'
 import { fabric } from 'fabric'
 import { useEffect, useState } from 'react'
 import { v4 as uuidv4 } from 'uuid'
 
 export default function Roof() {
-  const [canvas, setCanvas] = useState<fabric.Canvas>()
-  const [isLocked, setIsLocked] = useState<boolean>(false)
-  const [history, setHistory] = useState<fabric.Object[] | undefined>([])
-
-  useEffect(() => {
-    const c = new fabric.Canvas('canvas', {
-      height: 800,
-      width: 800,
-    })
-
-    // settings for all canvas in the app
-    fabric.Object.prototype.transparentCorners = false
-    fabric.Object.prototype.cornerColor = '#2BEBC8'
-    fabric.Object.prototype.cornerStyle = 'rect'
-    fabric.Object.prototype.cornerStrokeColor = '#2BEBC8'
-    fabric.Object.prototype.cornerSize = 6
-
-    setCanvas(c)
-    return () => {
-      c.dispose()
-    }
-  }, [])
-
-  useEffect(() => {
-    if (canvas) {
-      initialize()
-    }
-  }, [canvas])
-
-  const initialize = () => {
-    canvas?.clear()
-    /**
-     * 눈금 그리기
-     */
-    const width = canvas?.getWidth()
-    const height = canvas?.getHeight()
-
-    let startX = 0
-
-    let startY = 0
-
-    while (startX <= width!) {
-      startX += 10
-      const verticalLine = new fabric.Line([startX, 0, startX, height!], {
-        name: uuidv4(),
-        stroke: 'black',
-        strokeWidth: 1,
-        selectable: false,
-        opacity: 0.5,
-      })
-
-      createFigure(verticalLine)
-    }
-
-    while (startY <= height!) {
-      startY += 10
-      const verticalLine = new fabric.Line([0, startY, width!, startY], {
-        name: uuidv4(),
-        stroke: 'black',
-        strokeWidth: 1,
-        selectable: false,
-        opacity: 0.5,
-      })
-
-      createFigure(verticalLine)
-    }
-  }
+  const [
+    canvas,
+    createFigure,
+    handleUndo,
+    handleRedo,
+    handleClear,
+    handleCopy,
+    handleDelete,
+  ]: [
+    fabric.Canvas | undefined,
+    Function,
+    Function,
+    Function,
+    Function,
+    Function,
+    Function,
+  ] = useCanvas('canvas')
 
   const addRect = () => {
     const rect = new fabric.Rect({
@@ -124,41 +75,6 @@ export default function Roof() {
     createFigure(triangle)
   }
 
-  const createFigure = (figure: fabric.Object) => {
-    canvas?.add(figure)
-    canvas?.setActiveObject(figure)
-    canvas?.requestRenderAll()
-  }
-
-  const handleClear = () => {
-    canvas?.clear()
-    initialize()
-  }
-
-  const handleCopyFigure = () => {
-    const targetObj = canvas?.getActiveObject()
-    if (!targetObj) {
-      alert('복사할 대상을 선택해주세요.')
-      return
-    }
-
-    targetObj.clone((cloned: fabric.Object) => {
-      cloned.left! += 10
-      cloned.top! += 10
-      canvas?.add(cloned)
-      canvas?.setActiveObject(cloned)
-    })
-  }
-
-  const handleDelete = () => {
-    const targetObj = canvas?.getActiveObject()
-    if (!targetObj) {
-      alert('삭제할 대상을 선택해주세요.')
-      return
-    }
-    canvas?.remove(targetObj)
-  }
-
   const handleSave = () => {
     const objects = canvas?.getObjects()
 
@@ -168,7 +84,7 @@ export default function Roof() {
     }
     const jsonStr = JSON.stringify(canvas)
     localStorage.setItem('canvas', jsonStr)
-    initialize()
+    handleClear()
   }
 
   const handlePaste = () => {
@@ -187,64 +103,6 @@ export default function Roof() {
   const randomColor = () => {
     return '#' + Math.round(Math.random() * 0xffffff).toString(16)
   }
-
-  const handleUndo = () => {
-    if (canvas) {
-      if (canvas._objects.length > 0) {
-        const poppedObject = canvas._objects.pop()
-        setHistory((prev: fabric.Object[] | undefined) => {
-          if (prev === undefined) {
-            return poppedObject ? [poppedObject] : []
-          }
-          return poppedObject ? [...prev, poppedObject] : prev
-        })
-        canvas.renderAll()
-      }
-    }
-  }
-
-  const handleRedo = () => {
-    if (canvas && history) {
-      if (history.length > 0) {
-        setIsLocked(true)
-        canvas.add(history[history.length - 1])
-        const newHistory = history.slice(0, -1)
-        setHistory(newHistory)
-      }
-    }
-  }
-
-  const onChange = (e: fabric.IEvent) => {
-    const target = e.target
-    if (target) {
-      settleDown(target)
-    }
-
-    if (!isLocked) {
-      setHistory([])
-    }
-    setIsLocked(false)
-  }
-
-  const settleDown = (figure: fabric.Object) => {
-    /**
-     * 10단위로 움직이도록
-     */
-    const left = Math.round(figure?.left! / 10) * 10
-    const top = Math.round(figure?.top! / 10) * 10
-
-    figure?.set({ left: left, top: top })
-  }
-
-  useEffect(() => {
-    if (!canvas) {
-      return
-    }
-
-    canvas.on('object:added', onChange)
-    canvas.on('object:modified', onChange)
-    canvas.on('object:removed', onChange)
-  }, [canvas])
 
   return (
     <>
@@ -275,7 +133,7 @@ export default function Roof() {
         </button>
         <button
           className="w-30 mx-2 p-2 rounded bg-black text-white"
-          onClick={handleCopyFigure}
+          onClick={handleCopy}
         >
           COPY FIGURE
         </button>
